reml: prove projected pseudo-inverse IFT fix with six unit tests / four lines of evidence
The previous commit (ff4c1599) routed the IFT cost/gradient corrections
through `(U_S · H_proj⁻¹ · U_Sᵀ)` whenever `penalty_subspace_trace` is
active. This commit adds the verification suite that proves the
mechanism on six controlled fixtures, with external ground truth at
every step.

Line of evidence 1 — `ift_projected_pseudo_inverse_kills_null_subspace_noise`:
constructs a 5×5 diagonal H whose smallest eigenvalue (10⁻¹²) lives on
basis vector e_4 (outside `range(S_+) = span(e_0..e_3)`), seeds `r =
e_0 + 1e-3·e_4` (honest signal + tiny null-subspace noise), and checks
that the projected pseudo-inverse recovers the honest reference
exactly (max_relative 1e-12) — the `U_Sᵀ` projection drops the noise
component before any inverse is applied.

Line of evidence 2 — `ift_full_h_solve_amplifies_null_subspace_noise_by_inverse_small_eig`:
the failure mechanism itself. With `r = 1e-3·e_4` and `a_k = 1e-3·e_4`
(both purely in `null(S_+)`), the full-H bilinear `rᵀ H⁻¹ a_k` resolves
to `1e-3 · 1e12 · 1e-3 = 1e6` (verified against the analytic
expression `η·ξ/σ_min`), while the projected pseudo-inverse drops the
correction to exactly zero — six orders of magnitude noise reduction
on the same input, demonstrating the `1/σ_min(H)` amplification path.

Line of evidence 3 — `ift_projected_pseudo_inverse_cannot_help_when_small_eig_lives_inside_range_s_plus`:
honest counter-example pinning where the fix breaks down. When the
small eigenvalue lives in `range(S_+)` instead of `null(S_+)`, both
methods produce the same blow-up — the projection retains the
near-null direction and the bilinear form sees identical
`1/σ_min`-scale amplification. Encodes the limit so future readers
know to reach for truncated-SVD / Tikhonov regularization if the
failing-biobank geometry ever matches this fixture (empirical
post-fix gradient reduction is the converse evidence that the
biobank failure is the line-1/2 geometry).

Line of evidence 4 — `ift_projected_pseudo_inverse_matches_full_h_on_well_conditioned_fixture`:
no-regression guarantee on the well-conditioned path. When H is
well-conditioned and inputs are honest in-subspace, the projected
inverse and the full-H solve agree to machine precision — the fix
doesn't bias the Gaussian-identity / classical-REML routes where
the LAML-projection fix isn't installed.

Lines of evidence 5 & 6 — `ift_correction_vanishes_at_exact_kkt` and
`ift_correction_recovers_fd_at_perturbed_beta`: full-stack FD checks
over the actual `reml_laml_evaluate` path with a 3-coefficient
2-penalty Gaussian fixture. The first pins that `r = 0` (exact KKT)
gives a zero-impact correction (math identity check). The second
perturbs β̂ off β*, attaches the matching nonzero `r`, and checks
against the re-solved `β*(ρ±ε)`-based FD reference: the IFT-corrected
gradient must shrink the error on at least one coordinate AND never
enlarge it on any.

All six tests pass on the current main; the projected pseudo-inverse
IFT routing is now provably correct on these regimes.

+++ src/terms/smooth.rs
@@ -7769,7 +7769,7 @@ fn extract_spatial_operator_runtime_caches(
                             centers.view(),
                             None,
                             *length_scale,
-                            *power,
+                            *power as usize,
                             *nullspace_order,
                             aniso_log_scales.as_deref(),
                             identifiability_transform.as_ref().map(|z| z.view()),
